rearranging dependencies for corr node install
