--- realtime/tmpscript.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-rt="true"
-if [ "$rt" = true ];
-then
-	echo $rt
-else
-	echo nope
-fi
