+++ scripts/hdrtestnew.txt
@@ -22,10 +22,10 @@ NPOL 2
 
 OBS_OFFSET 0
 
-UTC_START 2025-06-16-15:07:32
+UTC_START 2025-06-16-18:33:05
 TSAMP 0.00125
 RESOLUTION 4
 
-MJD 60843.1302314814 
+MJD 60843.2729745370 
 SB 1 
 DEC 71.6
